Fix kernel indexing and review feedback

+++ src/include/msccl/msccl_struct.h
@@ -144,8 +144,8 @@ struct mscclAlgo {
   bool inPlace;
   // Whether this algorithm is suitable for out-of-place.
   bool outOfPlace;
-  // Keep a bit mask of used types
-  uint64_t typeMask;
+  // Keep a bit mask of used types (max 8 at present)
+  uint8_t typeMask;
 };
 
 enum mscclGroupStatus {
